Working version of JSON schema detection
